+++ PMFightAcademy.Admin/Contract/CoachContract.cs
@@ -1,7 +1,4 @@
-using System;
 using System.ComponentModel.DataAnnotations;
-using System.ComponentModel.DataAnnotations.Schema;
-using System.Text.Json.Serialization;
 
 namespace PMFightAcademy.Admin.Contract
 {
